Fix dtype issue

+++ graphite/data/dataset.py
@@ -457,7 +457,7 @@ class Dataset():
                             x_aug_data = utils.prepare_image_batch(x_aug_data, self.image_shape)
                             y_aug_data = utils.prepare_text_batch(y_aug_data, self.tokenizer.lexical_shape)
 
-                        w_data = np.array([data['writer'] for data in batch], dtype=np.int32)
+                        w_data = np.array([data['writer'] for data in batch], dtype=np.int64)
                         x_data = (x_data, y_data, x_aug_data, y_aug_data, w_data)
 
                     elif 'recognition' in self.workflow:
